byte ti int

+++ rabcrClient/ClothesTypes.cs
@@ -57,7 +57,7 @@ namespace rabcrClient {
         public bool ShinkHair;
         public Color Color;
     }
-    public enum HandClothSize : byte{
+    public enum HandClothSize : int{
         None=0,
         Half=18/2,
         NearlyFull=18-4,
